requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.45.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
idna==3.2
kombu==5.1.0
mysqlclient @ file:///C:/Users/user/Desktop/mysqlclient-1.4.6-cp38-cp38-win32.whl
oauthlib==3.1.1
prompt-toolkit==3.0.19
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
